v2: move parsing functions to their own package.
