+++ v2/internal/pkg/parser/parse_test.go
@@ -1,4 +1,4 @@
-package common
+package parser
 
 import (
 	"os"
